Changed the Circle::PointIsOn() method to allow two 'accuracy' distances because each of the two points being compared can be 'accuracy' distance from its real location.  This repairs a fault when trying to re-interpret points to arcs on the way 'back out of' the libArea code.
Added various changes to enable building on a Windows platform.

Changed the 'dirent' method of directory listing to the QDir class so that it would compile on Windows.

Changed various 'qstring.h' references to QString so that it would stop giving the 'deprecated' warning on the Windows build.

Added some of the QString::fromAscii() calls to literal strings.