Changed lots of printf() calls to qDebug(), qWarning() etc.
Changed references to PreCompiled.h so that none of them included a path to the header file within the '#include' line.  Instead I've added an extra path to the CMakeLists.txt files so that the correct PreCompiled.h file was chosen based on the project rather than the source file that refers to it.  This allows header files to be included in both the CAM and CAMGUI projects with each one selecting the PreCompiled.h file for its own project.  Without this, we end up with conflicting __declspec(dllimport)/dllexport definitions.

Removed TPGSettings.cpp from the CamGui project as it's defined within the Cam.dll file and is, therefore, multiply defined if it's also included in the CamGui project.