Changed references to PreCompiled.h from the 'double quotes' to the angle brackets form of inclusion in an attempt to stop it looking in the 'current' directory before looking in the INCLUDE paths given to the compiler.
Removed the FCAppCam pre-processor from the Windows build of the CamGui project so that it could be used to indicate whether we're compiling the 'Cam' library as opposed to the 'CamGui' or any of the plugins projects.  This allows some extra macros to be implemented in each of the two PreCompiled.h files to specify the correct dllimport/dllexport values based on the project rather than on which file includes which PreCompiled.h file.

Moved lots of method implementations out of the header files into the CPP files so that we don't end up with duplicate implementations when the CAM library is linked with the CamGui and/or Plugin libraries on the Windows build.