Added spindle speed and feedrate settings to the example TPG.
Changed the way the Python scripts handle feedrate and spindle speed GCode.  Rather than accumulate text to add to other GCode, it simply prints the feedrate and spindle speed values on their own lines.  This is much more simple.

Removed the 'modal' sequences from the Python scripts in an attempt to simplify them considerably.  If someone wants to produce modal GCode then my preference is for them to override the various methods in a machine-specific post processor rather than in the base classes (iso.py).

Removed the option to specify the i, j and k arguments in the arc() methods with absolute values rather than relative values.  Again, this just made the Python routines more complicated than they needed to be.

Removed the set_machine_coordinates() method as the right place for this is as a parameter to the rapid() and feed() methods instead.  i.e. if this flag is true then a G53 sequence is added before the X,Y,Z values are printed.