Fixed up some formatting issues (too many spaces) in the Python scripts.
Fixed some constructors and clone() methods so that ALL member variables are correctly initialized no matter which constructor was used.  This fixes the error where the CppExampleTPG could not read the 'settings' that had previously been added.  i.e. there was an invalid pointer due to it not having been initialized during a clone() call.