Removed some unnecessary braces in ToolPath.cpp
Commented out the example code in CppExampleTPG.cpp that makes use of the libArea module.  I can't get
the libarea module to load in Python yet.  I've left the example in there as comments as it shows how
the Cam::Paths object can be used to generate ToolPath rows of Python code to represent the graphics.  This
leads into the ability to feed the graphics into existing Python scripts for toolpath generation.